refactor(providers): extract PrideProvider with multi-protocol fallback
Moved PRIDE-specific logic (V3 API listing, multi-protocol batch
downloader with aspera/s3/ftp/globus fallback, private-dataset path,
submitter helpers, Globus/S3 per-protocol downloaders, legacy
single-connection FTP) from Files into providers/pride.py as
PrideProvider(Provider). ~510 LOC moved out of files.py.

Files keeps shim methods for every patched helper
(_batch_download_by_protocol, _download_with_fallback,
_globus_download_one, download_files_from_globus, download_files_from_s3,
download_files_from_ftp, download_private_file_name,
get_ascp_binary, save_checksum_file, stream_all_files_by_project,
stream_all_files_metadata, get_submitted_file_path_prefix,
_protocol_sequence, download_files).

PRIDE class constants (V3_API_BASE_URL, API_BASE_URL, API_PRIVATE_URL,
PRIDE_ARCHIVE_FTP, *_URL_PREFIX, S3_URL, S3_BUCKET, PROTOCOL_ORDER)
remain on Files as re-exports.

PrideProvider's internal calls to patch-sensitive helpers go through
Files.X (lazy import) so existing test patches in
test_download_resilience.py keep working without modification.

is_direct_download_accession updated to exclude PRIDE (returns True
only for MSV/JPST/IPX) now that PrideProvider is registered too.

All 8 patch-sensitive tests in test_download_resilience.py pass.
Full suite green at 67 passed, 4 skipped.
